backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
requests
Pillow